Added feature. Fixed bug.
Added a feature to prompt the user for the filename to load when the
program starts. It will check if the filename is valid, and if not, will
allow the user to try again.

Fixed bug in "I'm feeling lucky" option where the 'yes' choice would
always be selected even if the user entered no.